# Reduces allocations in `GetParallelChunks`.