I was seeing a lot of allocation in this code path and it turns out snappy reader is expensive to create.

This uses a pool in GetParallelChunks that way snappy reader are re-used.

benchmp:

```
❯ benchcmp before.txt after.txt
benchmark                         old ns/op     new ns/op     delta
BenchmarkGetParallelChunks-16     18147888      1483198       -91.83%

benchmark                         old allocs     new allocs     delta
BenchmarkGetParallelChunks-16     4214           9              -99.79%

benchmark                         old bytes     new bytes     delta
BenchmarkGetParallelChunks-16     146576089     127882        -99.91%
```
Signed-off-by: Cyril Tovena <cyril.tovena@gmail.com>